Split some general coveo helpers out of browse card and into coveo-exl-pipeline-helpers
