+++ blocks/bookmarks/bookmarks.js
@@ -3,11 +3,11 @@ import { buildCard } from '../../scripts/browse-card/browse-card.js';
 import { htmlToElement, fetchLanguagePlaceholders, getPathDetails } from '../../scripts/scripts.js';
 import { defaultProfileClient } from '../../scripts/auth/profile.js';
 import { fetchArticleByID } from '../../scripts/data-service/article-data-service.js';
-import { CONTENT_TYPES } from '../../scripts/browse-card/browse-cards-constants.js';
 import BuildPlaceholder from '../../scripts/browse-card/browse-card-placeholder.js';
 import { bookmarksEventEmitter } from '../../scripts/events.js';
 import { getCardData, convertToTitleCase } from '../../scripts/browse-card/browse-card-utils.js';
 import Pagination from '../../scripts/pagination/pagination.js';
+import { CONTENT_TYPES } from '../../scripts/data-service/coveo/coveo-exl-pipeline-constants.js';
 
 const BOOKMARKS_BY_PG_CONFIG = {};
 const CARDS_MODEL = {};
