+++ scripts/browse-card/browse-cards-constants.js
@@ -10,49 +10,6 @@ try {
   console.error('Error fetching placeholders:', err);
 }
 
-export const CONTENT_TYPES = Object.freeze({
-  COURSE: {
-    MAPPING_KEY: 'course',
-    LABEL: placeholders.browseCardCourseLabel || 'Course',
-  },
-  TUTORIAL: {
-    MAPPING_KEY: 'tutorial',
-    LABEL: placeholders.browseCardTutorialLabel || 'Tutorial',
-  },
-  DOCUMENTATION: {
-    MAPPING_KEY: 'documentation',
-    LABEL: placeholders.browseCardDocumentationLabel || 'Documentation',
-  },
-  TROUBLESHOOTING: {
-    MAPPING_KEY: 'troubleshooting',
-    LABEL: placeholders.browseCardTroubleshootingLabel || 'Troubleshooting',
-  },
-  EVENT: {
-    MAPPING_KEY: 'event',
-    LABEL: placeholders.browseCardEventLabel || 'On-Demand Event',
-  },
-  COMMUNITY: {
-    MAPPING_KEY: 'community',
-    LABEL: placeholders.browseCardCommunityLabel || 'Community',
-  },
-  CERTIFICATION: {
-    MAPPING_KEY: 'certification',
-    LABEL: placeholders.browseCardCertificationLabel || 'Certification',
-  },
-  LIVE_EVENT: {
-    MAPPING_KEY: 'live-event',
-    LABEL: placeholders.browseCardLiveEventLabel || 'Live Event',
-  },
-  INSTRUCTOR_LED: {
-    MAPPING_KEY: 'instructor-led-training',
-    LABEL: placeholders.browseCardInstructorLedLabel || 'Instructor-Led',
-  },
-  PERSPECTIVE: {
-    MAPPING_KEY: 'perspective',
-    LABEL: placeholders.browseCardPerspectiveLabel || 'Perspective',
-  },
-});
-
 export const COVEO_SORT_OPTIONS = Object.freeze({
   RELEVANCE: 'relevancy',
   MOST_RECENT: 'date descending',
@@ -72,25 +29,6 @@ export const ROLE_OPTIONS = Object.freeze({
   LEADER: 'Leader',
   USER: 'User',
 });
-
-export const COMMUNITY_SEARCH_FACET = Object.freeze([
-  {
-    value: 'Questions',
-    state: 'selected',
-  },
-  {
-    value: 'Discussions',
-    state: 'selected',
-  },
-  {
-    value: 'Ideas',
-    state: 'selected',
-  },
-  {
-    value: 'Blogs',
-    state: 'selected',
-  },
-]);
 
 export const RECOMMENDED_COURSES_CONSTANTS = Object.freeze({
   IN_PROGRESS: {
