+++ scripts/browse-card/browse-cards-delegate.js
@@ -4,73 +4,16 @@ import ADLSDataService from '../data-service/adls-data-service.js';
 import BrowseCardsCoveoDataAdaptor from './browse-cards-coveo-data-adaptor.js';
 import BrowseCardsLiveEventsAdaptor from './browse-cards-live-events-adaptor.js';
 import BrowseCardsADLSAdaptor from './browse-cards-adls-adaptor.js';
-import { CONTENT_TYPES, COMMUNITY_SEARCH_FACET, RECOMMENDED_COURSES_CONSTANTS } from './browse-cards-constants.js';
+import { CONTENT_TYPES } from '../data-service/coveo/coveo-exl-pipeline-constants.js';
 import PathsDataService from '../data-service/paths-data-service.js';
-import { getConfig, getPathDetails } from '../scripts.js';
+import { URL_SPECIAL_CASE_LOCALES, getConfig, getPathDetails } from '../scripts.js';
+import { getExlPipelineDataSourceParams } from '../data-service/coveo/coveo-exl-pipeline-helpers.js';
+import { RECOMMENDED_COURSES_CONSTANTS } from './browse-cards-constants.js';
 
-const { coveoSearchResultsUrl, liveEventsUrl, adlsUrl, pathsUrl } = getConfig();
+const { liveEventsUrl, adlsUrl, pathsUrl } = getConfig();
 
 const { lang } = getPathDetails();
 
-// Most of these are copied from an existing call. I do not believe we need all of them, so this list could probably be pruned.
-const fieldsToInclude = [
-  '@foldingchild',
-  '@foldingcollection',
-  '@foldingparent',
-  'author',
-  'author_bio_page',
-  'author_name',
-  'author_type',
-  'authorname',
-  'authortype',
-  'collection',
-  'connectortype',
-  'contenttype',
-  'date',
-  'documenttype',
-  'el_author_type',
-  'el_contenttype',
-  'el_id',
-  'el_interactionstyle',
-  'el_kudo_status',
-  'el_lirank',
-  'el_product',
-  'el_rank_icon',
-  'el_reply_status',
-  'el_solution',
-  'el_solutions_authored',
-  'el_type',
-  'el_usergenerictext',
-  'el_version',
-  'el_view_status',
-  'exl_description',
-  'exl_thumbnail',
-  'filetype',
-  'id',
-  'language',
-  'liMessageLabels',
-  'liboardinteractionstyle',
-  'licommunityurl',
-  'lithreadhassolution',
-  'objecttype',
-  'outlookformacuri',
-  'outlookuri',
-  'permanentid',
-  'role',
-  'source',
-  'sourcetype',
-  'sysdocumenttype',
-  'type',
-  'urihash',
-  'video_url',
-];
-
-const locales = new Map([
-  ['es', 'es-ES'],
-  ['pt-br', 'pt-BR'],
-  ['zh-hans', 'zh-CN'],
-  ['zh-hant', 'zh-TW'],
-]);
 /**
  * @module BrowseCardsDelegate
  * @description A module that handles the delegation of fetching card data based on content types.
@@ -84,103 +27,13 @@ const BrowseCardsDelegate = (() => {
   let param = {};
 
   /**
-   * Constructs Coveo facet structure based on provided facets.
-   * @param {Array} facets - An array of facet objects.
-   * @returns {Array} Array of Coveo facet objects.
-   * @private
-   */
-  const constructCoveoFacet = (facets) => {
-    const facetsArray = facets.map((facet) => ({
-      facetId: `@${facet.id}`,
-      field: facet.id,
-      type: facet.type,
-      numberOfValues: facet.currentValues?.length || 2,
-      currentValues: facet.currentValues.map((value) => ({
-        value,
-        state: value === CONTENT_TYPES.COMMUNITY.MAPPING_KEY ? 'idle' : 'selected',
-        ...(value === CONTENT_TYPES.COMMUNITY.MAPPING_KEY ? { children: COMMUNITY_SEARCH_FACET } : []),
-      })),
-    }));
-    return facetsArray;
-  };
-
-  /**
-   * Constructs advanced query for Coveo data service based on query parameters.
-   * @returns {Object} Object containing the advanced query.
-   * @private
-   */
-  const constructCoveoAdvancedQuery = () => {
-    const featureQuery = param.feature
-      ? `(${param.feature.map((type) => `@el_features=="${type}"`).join(' OR ')})`
-      : '';
-    const contentTypeQuery = param.contentType
-      ? `AND (${param.contentType.map((type) => `@el_contenttype=="${type}"`).join(' OR ')})`
-      : '';
-    const productQuery = param.product
-      ? `AND (${param.product.map((type) => `@el_product=="${type}"`).join(' OR ')})`
-      : '';
-    const versionQuery = param.version
-      ? `AND (${param.version.map((type) => `@el_version=="${type}"`).join(' OR ')})`
-      : '';
-    const roleQuery = param.role ? `AND (${param.role.map((type) => `@el_role=="${type}"`).join(' OR ')})` : '';
-    const levelQuery = param.level ? `AND (${param.level.map((type) => `@el_level=="${type}"`).join(' OR ')})` : '';
-    const authorTypeQuery = param.authorType
-      ? `AND (${param.authorType.map((type) => `@author_type=="${type}"`).join(' OR ')})`
-      : '';
-    const query = `${featureQuery} ${contentTypeQuery} ${productQuery} ${versionQuery} ${roleQuery} ${levelQuery} ${authorTypeQuery}`;
-    return query;
-  };
-
-  /**
-   * Constructs advanced query for Coveo data service based on date array.
-   * @param {Array} dateCriteria - Array of date values.
-   * @returns {string} Advanced query string for date.
-   */
-  const contructDateAdvancedQuery = (dateCriteria) => `@date==(${dateCriteria.join(',')})`;
-
-  /**
    * Handles Coveo data service to fetch card data.
    * @returns {Array} Array of card data.
    * @throws {Error} Throws an error if an issue occurs during data fetching.
    * @private
    */
   const handleCoveoService = async () => {
-    const facets = [
-      ...(param.contentType
-        ? [
-            {
-              id: 'el_contenttype',
-              type: param.contentType[0] === CONTENT_TYPES.COMMUNITY.MAPPING_KEY ? 'hierarchical' : 'specific',
-              currentValues: param.contentType,
-            },
-          ]
-        : []),
-      ...(param.product ? [{ id: 'el_product', type: 'specific', currentValues: param.product }] : []),
-      ...(param.version ? [{ id: 'el_version', type: 'specific', currentValues: param.version }] : []),
-      ...(param.role ? [{ id: 'el_role', type: 'specific', currentValues: param.role }] : []),
-      ...(param.authorType ? [{ id: 'author_type', type: 'specific', currentValues: param.authorType }] : []),
-      ...(param.level ? [{ id: 'el_level', type: 'specific', currentValues: param.level }] : []),
-    ];
-
-    const dataSource = {
-      url: coveoSearchResultsUrl,
-      param: {
-        locale: locales.get(document.querySelector('html').lang) || document.querySelector('html').lang || 'en',
-        searchHub: `Experience League Learning Hub`,
-        numberOfResults: param.noOfResults,
-        excerptLength: 200,
-        sortCriteria: param.sortCriteria,
-        context: { entitlements: {}, role: {}, interests: {}, industryInterests: {} },
-        filterField: '@foldingcollection',
-        parentField: '@foldingchild',
-        childField: '@foldingparent',
-        ...(param.q && !param.feature ? { q: param.q } : ''),
-        ...(param.dateCriteria && !param.feature ? { aq: contructDateAdvancedQuery(param.dateCriteria) } : ''),
-        ...(!param.feature ? { facets: constructCoveoFacet(facets) } : ''),
-        ...(param.feature ? { aq: constructCoveoAdvancedQuery() } : ''),
-        fieldsToInclude,
-      },
-    };
+    const dataSource = getExlPipelineDataSourceParams(param);
     const coveoService = new CoveoDataService(dataSource);
     const cardData = await coveoService.fetchDataFromSource();
     if (!cardData) {
@@ -260,7 +113,7 @@ const BrowseCardsDelegate = (() => {
    * @private
    */
   const constructPathsSearchParams = () => {
-    const languageParam = locales.get(lang) || lang;
+    const languageParam = URL_SPECIAL_CASE_LOCALES.get(lang) || lang;
     const urlSearchParams = new URLSearchParams();
     urlSearchParams.append('page_size', '200');
     urlSearchParams.append('sort', 'Order,Solution,ID');
