+++ scripts/browse-card/browse-cards-paths-adaptor.js
@@ -1,6 +1,7 @@
 import browseCardDataModel from '../data-model/browse-cards-model.js';
-import { CONTENT_TYPES, RECOMMENDED_COURSES_CONSTANTS } from './browse-cards-constants.js';
+import {RECOMMENDED_COURSES_CONSTANTS } from './browse-cards-constants.js';
 import { fetchLanguagePlaceholders, getConfig } from '../scripts.js';
+import { CONTENT_TYPES } from '../data-service/coveo/coveo-exl-pipeline-constants.js';
 
 const { recommendedCoursesUrl, prodAssetsCdnOrigin } = getConfig();
 /**
